requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
cycler==0.11.0
defusedxml==0.7.1
Django==4.2.5
django-compat==1.0.15
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
fonttools==4.42.1
idna==3.4
importlib-resources==6.0.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.4
zipp==3.16.2
